Simplified Entites
Change /weaponname/shared.lua to /weaponname.lua
