+++ entities/entities/info_hidden_spawn.lua
